Причесано все.

+++ NumbersNN/Network.h
@@ -10,7 +10,7 @@ public:
 	Network(const std::vector<int> om0s = std::vector<int>(10, 10), double eps = 0.5);
 
 	void setOm0s(const std::vector<int> om0s);
-	std::vector<int> om0s() const;
+	std::vector<double> om0s() const;
 
 	void setEps(double eps);
 	double eps() const;
